+++ CHANGELOG.md
@@ -2,5 +2,6 @@
   - New Feature: Add `kkplot.xticks` and `kkplot.yticks` as wrappers around `kkplot.ticks` for the corresponding axis.
   - Change: `kkconfig.local` now provides a `load_settings` function instead of directly loading the settings.
   - New Feature: Add `kkcoffea.HistProcessor.weight` to return the event weights given `weight_column`.
+  - Change: Restructure `kkcoffea.Samples.SampleManager` sample input format to allow settings passed in sample defs.
 
 0.0.1: Initial Release